+++ muschart/src/main/webapp/static/js/angular/factory/track.factory.js
@@ -1,7 +1,10 @@
 'use strict';
-app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', function($http, MESSAGE, REST) {
+app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', 'ValidatorService', function($http, MESSAGE, REST, ValidatorService) {
 
 	function createTrack(name, song, cover, video, release, artists, units, genres, callback) {
+		if (!ValidatorService.allNotEmpty(callback, name, song, cover, video, release, artists, units, genres)) {
+			return;
+		}
 		$http.post(REST.TRACKS + '/create/' + name + '/' + song + '/' + cover + '/' + video + '/' + release + '/' + artists + '/' + units + '/' + genres + REST.JSON_EXT)
 		.success(function(response) {
 			var data = {success: true, data: response, message: MESSAGE.CREATING_TRACK_SUCCESS};
@@ -14,6 +17,9 @@ app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', function($http, MESSAGE
 	}
 
 	function deleteTrack(id, callback) {
+		if (!ValidatorService.allNotEmpty(callback, id)) {
+			return;
+		}
 		$http.delete(REST.TRACKS + '/delete/' + id + REST.JSON_EXT)
 		.success(function(response) {
 			response = {success: true, message: MESSAGE.GELETING_TRACK_SUCCESS};
@@ -26,6 +32,9 @@ app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', function($http, MESSAGE
 	}
 
 	function getTrackById(id, callback) {
+		if (!ValidatorService.allNotEmpty(callback, id)) {
+			return;
+		}
 		$http.get(REST.TRACKS + '/' + id + REST.JSON_EXT)
 		.success(function(response) {
 			var data = {success: true, data: response};
@@ -38,6 +47,9 @@ app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', function($http, MESSAGE
 	}
 
 	function getTracks(sort, order, page, callback) {
+		if (!ValidatorService.allNotEmpty(callback, sort, order, page)) {
+			return;
+		}
 		$http.get(REST.TRACKS + '/' + sort + '/' + order + '/' + page + REST.JSON_EXT)
 		.success(function(response) {
 			var data = {success: true, data: response};
@@ -50,6 +62,9 @@ app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', function($http, MESSAGE
 	}
 
 	function getEntityTracks(entity, entityId, sort, order, page, callback) {
+		if (!ValidatorService.allNotEmpty(callback, entity, entityId, sort, order, page)) {
+			return;
+		}
 		$http.get(REST.TRACKS + '/' + entity + '/' + entityId + '/' + sort + '/' + order + '/' + page + REST.JSON_EXT)
 		.success(function(response) {
 			var data = {success: true, data: response};
@@ -62,6 +77,9 @@ app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', function($http, MESSAGE
 	}
 
 	function getUserTracks(sort, order, page, callback) {
+		if (!ValidatorService.allNotEmpty(callback, sort, order, page)) {
+			return;
+		}
 		$http.get(REST.TRACKS + '/user/' + sort + '/' + order + '/' + page + REST.JSON_EXT)
 		.success(function(response) {
 			var data = {success: true, data: response};
@@ -98,6 +116,9 @@ app.factory('TrackFactory', ['$http', 'MESSAGE', 'REST', function($http, MESSAGE
 	}
 
 	function getEntityPagesCount(entity, entityId, callback) {
+		if (!ValidatorService.allNotEmpty(callback, entity, entityId)) {
+			return;
+		}
 		$http.get(REST.TRACKS + '/' + entity + '/' + entityId + '/pages_count' + REST.JSON_EXT)
 		.success(function(response) {
 			var data = {success: true, data: response};
